fix: YAML-to-string broken conversions
It looks like I left this out of the previous PR.

This replaces the base `ruamel.yaml` requirement with the
string extension `ruamel.yaml.string` (which will pull in
the base `ruamel.yaml` package).  The string extension adds
the `dumps` method to serialise to a Python string, which is
needed with the new way of doing things.

`black` has also removed some blank lines, apparently.